refactor(filemanager): group resolvers in separate package
